protobuf: Properly detect protobuf messages as binary
Add testcases for the proto definition, text proto and the
resulting binary. We face the problem that with just 500
bytes we will not be able to fully parse all protobufs but
do our best.

Introduce a reader and start parsing messages. We will bail
when failing to parse a message before having reached the
end of the message.